+++ src/libromdata/Console/WiiSettingTxt.hpp
@@ -1,0 +1,18 @@
+/***************************************************************************
+ * ROM Properties Page shell extension. (libromdata)                       *
+ * WiiSettingTxt.hpp: Nintendo Wii setting.txt file reader.                *
+ *                                                                         *
+ * Copyright (c) 2016-2025 by David Korth.                                 *
+ * SPDX-License-Identifier: GPL-2.0-or-later                               *
+ ***************************************************************************/
+
+#pragma once
+
+#include "librpbase/RomData.hpp"
+
+namespace LibRomData {
+
+ROMDATA_DECL_BEGIN(WiiSettingTxt)
+ROMDATA_DECL_END()
+
+}