+++ src/librpbase/RomData.hpp
@@ -211,6 +211,7 @@ public:
 		MetadataFile,		// Metadata File
 		PatchFile,		// Patch File
 		Ticket,			// Ticket file (e.g. Wii/Wii U .tik)
+		ConfigurationFile,	// Configuration File
 
 		Max
 	};